Made a new DTO class for transforming the string we get from the python input side, into a json obeject we can use to update the state of the Board
